TSL bug fix
